+++ CMakeLists.txt
@@ -62,6 +62,7 @@ option(BUILD_VINEYARD_MALLOC_OVERRIDE "Using the client-side malloc to override 
 option(BUILD_VINEYARD_FUSE "Enable vineyard's fuse support" OFF)
 option(BUILD_VINEYARD_FUSE_PARQUET "Enable vineyard's fuse parquet support" OFF)
 option(BUILD_VINEYARD_HOSSEINMOEIN_DATAFRAME "Enable hosseinmoein dataframe support" OFF)
+option(BUILD_VINEYARD_KV_STATE_CACHE "Enable kv-state cache support" ON)
 
 option(BUILD_VINEYARD_TESTS "Generate make targets for vineyard tests" ON)
 option(BUILD_VINEYARD_TESTS_ALL "Include make targets for vineyard tests to ALL" OFF)
@@ -932,10 +933,10 @@ if(BUILD_VINEYARD_HOSSEINMOEIN_DATAFRAME)
     list(APPEND VINEYARD_INSTALL_LIBS vineyard_hosseinmoein_dataframe)
 endif()
 
-# if(BUILD_VINEYARD_KVSTATE_CACHE)
-add_subdirectory(modules/kv-state-cache)
-list(APPEND VINEYARD_INSTALL_LIBS vineyard_kv_state_cache)
-# endif()
+if(BUILD_VINEYARD_KV_STATE_CACHE)
+    add_subdirectory(modules/kv-state-cache)
+    list(APPEND VINEYARD_INSTALL_LIBS vineyard_kv_state_cache)
+endif()
 
 if(BUILD_VINEYARD_TESTS)
     add_subdirectory(test)
